+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to the Imperative package will be documented in this file.
 - BugFix: Fixed `config import` and `config init` behaving incorrectly when config JSON exists in higher level directory. [zowe/zowe-cli#1218](https://github.com/zowe/zowe-cli/issues/1218)
 - BugFix: Fixed `config import` command not failing when positional argument "location" is missing.
 
+## `5.0.0-next.202112012301`
+
+- Enhancement: Changed CLI prompt input to be hidden for properties designated as secure in team config. [zowe/zowe-cli#1106](https://github.com/zowe/zowe-cli/issues/1106)
+- BugFix: Improved error message when Keytar module fails to load. [#27](https://github.com/zowe/imperative/issues/27)
+- **Next Breaking**
+    - Removed the `ConfigProfiles.load` API method. Use the methods `ConfigLayers.find` and `ConfigSecure.securePropsForProfile` instead. [#568](https://github.com/zowe/imperative/issues/568)
+
 ## `5.0.0-next.202111301806`
 
 - Enhancement: Added a utility function to get basic system architecture and platform info